pyproject.toml
[tool.poetry]
name = "trazo-back"
version = "0.1.0"
description = ""
authors = ["Oscar Muñoz <osmunogue@gmail.com>"]
readme = "README.md"
packages = [
    { include = "backend" },
]


[tool.poetry.dependencies]
python = "^3.9"
asgiref = "3.6.0"
certifi = "2022.12.7"
cffi = "1.17.1"
charset-normalizer = "2.1.1"
click = "8.1.3"
cryptography = "38.0.4"
defusedxml = "0.7.1"
dj-database-url = "1.2.0"
django = "4.1.4"
django-allauth = "0.51.0"
django-cors-headers = "3.13.0"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.2.2"
gunicorn = "20.1.0"
idna = "3.4"
mypy-extensions = "0.4.3"
oauthlib = "3.2.2"
pathspec = "0.10.3"
pillow = "11.0.0"
platformdirs = "2.6.2"
psycopg2-binary = "2.9.10"
pycparser = "2.21"
pyjwt = "2.6.0"
python-dotenv = "0.21.0"
python3-openid = "3.2.0"
pytz = "2022.7"
qrcode = "7.3.1"
requests = "2.28.1"
requests-oauthlib = "1.3.1"
sqlparse = "0.4.3"
tomli = "2.0.1"
typing-extensions = "^4.5.0"
urllib3 = "1.26.13"
whitenoise = "6.2.0"
python-decouple = "^3.7"
geoip2 = "^4.7.0"
ipython = "^8.14.0"
boto3 = "^1.28.57"
django-storages = "^1.14.1"
drf-nested-routers = "^0.93.4"
celery = "^5.3.6"
redis = "^5.0.1"
stripe = "^11.6.0"
django-celery-beat = "^2.5.0"
django-redis = "^5.4.0"
reportlab = "^4.4.1"
web3 = "^7.12.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-django = "^4.8.0"
black = "^25.1.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "backend.settings"
python_files = ["test_*.py", "*_test.py", "testing/python/*.py"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
